requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor-5==0.2.18
django-jazzmin==3.0.1
django-js-asset==2.2.0
pillow==11.0.0
sqlparse==0.5.1
